+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 ### Added
 
 - The `defstate` macro now declares the resulting var as `:dynamic`. This enables the use of `binding` in tests.
-- The function `deref?`, which lets you inspect the value of a state, without starting it.
+- The State object now implements `IPersistentState`, thus now supports Clojure's `peek`. This can be used to inspect the value, without starting it.
 
 ### Changed